Take into account the timeout of the last ten requests in multi-agent requests
