+++ cloud/blockstore/libs/storage/partition_common/get_device_for_range_companion.cpp
@@ -4,6 +4,8 @@
 #include <cloud/blockstore/libs/storage/core/config.h>
 #include <cloud/blockstore/libs/storage/core/forward_helpers.h>
 #include <cloud/blockstore/libs/storage/partition_nonrepl/config.h>
+#include <cloud/blockstore/libs/storage/partition_nonrepl/model/device_stats.h>
+
 #include <cloud/storage/core/libs/actors/helpers.h>
 
 namespace NCloud::NBlockStore::NStorage {
@@ -20,9 +22,11 @@ TGetDeviceForRangeCompanion::TGetDeviceForRangeCompanion(
 TGetDeviceForRangeCompanion::TGetDeviceForRangeCompanion(
         EAllowedOperation allowedOperation,
         TStorageConfigPtr config,
-        TNonreplicatedPartitionConfigPtr partConfig)
+        TNonreplicatedPartitionConfigPtr partConfig,
+        const TVector<TDeviceStat>* const deviceStats)
     : Config(std::move(config))
     , PartConfig(std::move(partConfig))
+    , DeviceStats(deviceStats)
     , AllowedOperation(allowedOperation)
 {}
 
@@ -86,9 +90,21 @@ void TGetDeviceForRangeCompanion::HandleGetDeviceForRange(
     auto response = std::make_unique<
         TEvNonreplPartitionPrivate::TEvGetDeviceForRangeResponse>();
 
+    if (DeviceStats) {
+        Y_DEBUG_ABORT_UNLESS(requests[0].DeviceIdx < DeviceStats->size());
+
+        const auto& deviceStat = (*DeviceStats)[requests[0].DeviceIdx];
+        if (deviceStat.DeviceStatus != TDeviceStat::EDeviceStatus::Ok) {
+            ReplyCanNotUseDirectCopy(ev, ctx);
+            return;
+        }
+        response->RequestTimeout = deviceStat.WorstRequestTime();
+    }
+
     response->Device = requests[0].Device;
     response->DeviceBlockRange = requests[0].DeviceBlockRange;
-    response->RequestTimeout = GetMinRequestTimeout();
+    response->RequestTimeout =
+        Max(response->RequestTimeout, GetMinRequestTimeout());
     response->PartConfig = PartConfig;
 
     NCloud::Reply(ctx, *ev, std::move(response));
