+++ cloud/blockstore/libs/storage/partition_nonrepl/model/ya.make
@@ -4,6 +4,7 @@ INCLUDE(${ARCADIA_ROOT}/cloud/storage/deny_ydb_dependency.inc)
 
 SRCS(
     changed_ranges_map.cpp
+    device_stats.cpp
     disjoint_range_set.cpp
     processing_blocks.cpp
     replica_actors.cpp
