+++ cloud/blockstore/libs/storage/partition_nonrepl/part_nonrepl_rdma_actor.h
@@ -63,7 +63,8 @@ private:
     TGetDeviceForRangeCompanion GetDeviceForRangeCompanion{
         TGetDeviceForRangeCompanion::EAllowedOperation::ReadWrite,
         Config,
-        PartConfig};
+        PartConfig,
+        nullptr};
 
     bool UpdateCountersScheduled = false;
     TPartitionDiskCountersPtr PartCounters;
